pyproject.toml
[project]
name = "smarttek-demo"
version = "0.1.0"
description = "SmartTek AB – drönbaserad fastighetsinspektion pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    "jinja2>=3.1.0",
    "weasyprint>=61.0",
    "pydantic>=2.5.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "shapely>=2.0.0",
    "pyproj>=3.6.0",
    "python-dotenv>=1.0.0",
    "piexif>=1.1.3",
]

[project.optional-dependencies]
geo = [
    "geopandas>=0.14.0",
    "rasterio>=1.3.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline", "api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
